% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_LongituRF.R
\name{MERF}
\alias{MERF}
\title{Title}
\usage{
MERF(
  X,
  Y,
  id,
  Z,
  iter = 100,
  mtry = ceiling(ncol(X)/3),
  ntree = 500,
  time,
  sto,
  delta = 0.001,
  n_threads = 1
)
}
\description{
Title
}
\keyword{internal}
