% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_APRF.R
\name{predict.APRF}
\alias{predict.APRF}
\title{Obtain predictions from a Asset Pricing Random Forest (APRF).}
\usage{
\method{predict}{APRF}(object, X, id, time, ...)
}
\arguments{
\item{object}{: a \code{APRF} output of train_APRF() function.}

\item{X}{[matrix]: data.frame of the fixed effects for the new observations to be predicted.}

\item{id}{[vector]: vector of the identifiers of the new observations to be predicted.}

\item{time}{[vector]: vector of the time measurements of the new observations to be predicted.}

\item{...}{: low levels arguments.}
}
\value{
vector of the predicted output for the new observations.
}
\description{
Obtain predictions from a Asset Pricing Random Forest (APRF).
}
\examples{
\donttest{}

}
